test: consolidate harness IPC contracts
